backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
sqlalchemy>=2.0,<3.0
psycopg[binary]>=3.2,<4.0
pydantic-settings>=2.4,<3.0
python-jose[cryptography]>=3.3,<4.0
passlib>=1.7,<2.0
alembic>=1.13,<2.0
apscheduler>=3.10,<4.0
httpx>=0.27,<1.0
paho-mqtt>=2.1,<3.0
pandas>=2.2,<3.0
numpy>=2.0,<3.0
scikit-learn>=1.5,<2.0
ephem>=4.1,<5.0
astropy>=6.1,<7.0
joblib>=1.4,<2.0
python-dotenv>=1.0,<2.0
opencv-python-headless>=4.10.0,<5.0

